refactor: Extract components into its own file
